Begin to wire up GTFS models into live data
This commit replaces a lot of the API-specific models, and instead
finds an appropriate model from the database to enhance with API-specific
data. There is some significant ugliness in dealing with the responses
from BusTracker and TrainTracker - their API doesn't return GTFS data
so we have to make highly educated guesses about which rows to pull out
of the database.


Former-commit-id: 6c99912b7c48a5a29bec61434a6b5abdfa7c3ac9