Continue integrating and refactoring
Added some methods around missing data in the GTFS feed, and collapsed
duplicate models. I also updated the SQLite database with a few missing
routes (and fired off an inquisitive email to the CTA), so we update
that in this commit as well. Hopefully we won't have to update the
DB much more...


Former-commit-id: d721ce152adf01b53e1ad5f37cddf0c339aea59e