A better find_active_run, and a new association
The CTA sometimes does this with trips:

start -> -> -> -> -> -> -> -> -> end
                                  |
start <- <- <- <- <- <- <- <- <- start

...and classify that as one "run". This is particularly the case
for L runs.

So, naturally, that broke the find_active_run method. This commit
updates it (and fixes tests) to ensure that we try to find the active
run for the *direction* of travel as well as the time of day.

Additionally, we associate shapes -> trips, which is important if
you're drawing maps. The data was available, I just overlooked the
relationship before.